@@ -2,6 +2,7 @@
     "name": "validator-services",
     "version": "2.4.9",
     "description": "Backend microservice provider for the IATI Validator on the Unified Platform",
+    "type": "module",
     "scripts": {
         "start": "func start --javascript --cors http://localhost:4200",
         "start:verbose": "func start --javascript --verbose",
@@ -10,18 +11,18 @@
     },
     "author": "IATI",
     "dependencies": {
-        "dotenv": "^16.0.1",
-        "node-fetch": "^2.6.7",
-        "pg": "^8.7.3",
+        "dotenv": "^16.0.3",
+        "node-fetch": "^3.2.10",
+        "pg": "^8.8.0",
         "pg-hstore": "^2.3.4"
     },
     "devDependencies": {
         "@prettier/plugin-xml": "^2.2.0",
-        "eslint": "^8.22.0",
+        "eslint": "^8.27.0",
         "eslint-config-airbnb-base": "^15.0.0",
         "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-import": "^2.26.0",
-        "husky": "^8.0.1",
+        "husky": "^8.0.2",
         "lint-staged": "^13.0.3",
         "prettier": "^2.7.1"
     },
