@@ -1,159 +1,207 @@
-const { Pool } = require('pg');
-const config = require('../config/config');
+import pg from 'pg';
+import config from '../config/config.js';
 
-module.exports = {
-    query: async (sql, values = null) => {
-        const pool = new Pool(config.PGCONFIG);
-        const result = await pool.query(sql, values);
-        await pool.end();
+const { Pool } = pg;
 
-        return result.rows;
-    },
+const query = async (sql, values = null) => {
+    const pool = new Pool(config.PGCONFIG);
+    const result = await pool.query(sql, values);
+    await pool.end();
 
-    getFirstRow: async (sql, values = null) => {
-        const pool = new Pool(config.PGCONFIG);
-        const result = await pool.query(sql, values);
-        await pool.end();
+    return result.rows;
+};
 
-        if (result.rows.length > 0) {
-            return result.rows[0];
-        }
-        return null;
-    },
+const getFirstRow = async (sql, values = null) => {
+    const pool = new Pool(config.PGCONFIG);
+    const result = await pool.query(sql, values);
+    await pool.end();
 
-    getReportForUrl: async (url) => {
-        const sql = `
-            SELECT val.document_hash as registry_hash, val.document_id as registry_id, val.document_url, val.valid, val.report
-            FROM public.document as doc
-            LEFT JOIN validation as val ON doc.validation=val.id
-            WHERE doc.url = $1;
-        `;
+    if (result.rows.length > 0) {
+        return result.rows[0];
+    }
+    return null;
+};
 
-        return module.exports.getFirstRow(sql, [url]);
-    },
-
-    getReportWithoutErrorsForUrl: async (url) => {
-        const sql = `
+const getReportForUrl = async (url) => {
+    const sql = `
             SELECT val.document_hash as registry_hash,
-                   val.document_id as registry_id, 
+                   val.document_id as registry_id,
+                   doc.name as registry_name,
                    val.document_url,
                    val.valid,
-                   (SELECT case when val.report is null then null else jsonb_build_object(
-                        'valid',val.report->'valid',
-                        'summary',val.report->'summary',
-                        'fileType',val.report->'fileType',
-                        'iatiVersion',val.report->'iatiVersion'
-                    ) end as report)
+                   val.report
             FROM public.document as doc
             LEFT JOIN validation as val ON doc.validation=val.id
             WHERE doc.url = $1;
         `;
 
-        return module.exports.getFirstRow(sql, [url]);
-    },
+    return getFirstRow(sql, [url]);
+};
 
-    getReportForTestfile: async (guid) => {
-        const sql = `
+const valReportSummaryOnly = `(SELECT case when val.report is null then null else jsonb_build_object(
+                                    'valid',val.report->'valid',
+                                    'summary',val.report->'summary',
+                                    'fileType',val.report->'fileType',
+                                    'iatiVersion',val.report->'iatiVersion'
+                                ) end as report)`;
+
+const getReportWithoutErrorsForUrl = async (url) => {
+    const sql = `
+            SELECT val.document_hash as registry_hash,
+                   val.document_id as registry_id, 
+                   doc.name as registry_name, 
+                   val.document_url,
+                   val.valid,
+                   ${valReportSummaryOnly}
+            FROM public.document as doc
+            LEFT JOIN validation as val ON doc.validation=val.id
+            WHERE doc.url = $1;
+        `;
+
+    return getFirstRow(sql, [url]);
+};
+
+const getReportForTestfile = async (guid) => {
+    const sql = `
             SELECT valid, report, filename, guid, session_id
             FROM adhoc_validation
             WHERE guid = $1
         `;
 
-        return module.exports.getFirstRow(sql, [guid]);
-    },
+    return getFirstRow(sql, [guid]);
+};
 
-    getReportForHash: async (hash) => {
-        const sql = `
-            SELECT document_hash as registry_hash, document_id as registry_id, document_url, valid, report
-            FROM validation
-            WHERE document_hash = $1
-            ORDER BY id DESC
-            LIMIT 1
-        `;
-        return module.exports.getFirstRow(sql, [hash]);
-    },
-
-    getReportWithoutErrorsForHash: async (hash) => {
-        const sql = `
-            SELECT document_hash as registry_hash, document_id as registry_id, document_url, valid, 
-                (SELECT case when report is null then null else jsonb_build_object(
-                        'valid',report->'valid',
-                        'summary',report->'summary',
-                        'fileType',report->'fileType',
-                        'iatiVersion',report->'iatiVersion'
-                    ) end as report)
-            FROM validation
-            WHERE document_hash = $1
-            ORDER BY id DESC
-            LIMIT 1
-        `;
-        return module.exports.getFirstRow(sql, [hash]);
-    },
-
-    getReportForId: async (id) => {
-        const sql = `
-            SELECT val.document_hash as registry_hash, val.document_id as registry_id, val.document_url, val.valid, val.report
-            FROM public.document as doc
-            LEFT JOIN validation as val ON doc.validation=val.id
-            WHERE doc.id = $1;
-        `;
-        return module.exports.getFirstRow(sql, [id]);
-    },
-
-    getReportWithoutErrorsForId: async (id) => {
-        const sql = `
+const getReportForHash = async (hash) => {
+    const sql = `
             SELECT val.document_hash as registry_hash,
-                   val.document_id as registry_id, 
+                   val.document_id as registry_id,
+                   doc.name as registry_name,
                    val.document_url,
                    val.valid,
-                   (SELECT case when val.report is null then null else jsonb_build_object(
-                        'valid',val.report->'valid',
-                        'summary',val.report->'summary',
-                        'fileType',val.report->'fileType',
-                        'iatiVersion',val.report->'iatiVersion'
-                    ) end as report)
+                   val.report
+            FROM public.document as doc
+            LEFT JOIN validation as val ON doc.validation=val.id
+            WHERE doc.hash = $1
+            ORDER BY val.id DESC
+            LIMIT 1;
+        `;
+    return getFirstRow(sql, [hash]);
+};
+
+const getReportWithoutErrorsForHash = async (hash) => {
+    const sql = `
+            SELECT val.document_hash as registry_hash, 
+                   val.document_id as registry_id, 
+                   doc.name as registry_name,
+                   val.document_url,
+                   val.valid, 
+                   ${valReportSummaryOnly}
+            FROM public.document as doc
+            LEFT JOIN validation as val ON doc.validation=val.id
+            WHERE doc.hash = $1
+            ORDER BY val.id DESC
+            LIMIT 1;
+        `;
+    return getFirstRow(sql, [hash]);
+};
+
+const getReportForId = async (id) => {
+    const sql = `
+            SELECT val.document_hash as registry_hash, 
+                   val.document_id as registry_id,
+                   doc.name as registry_name,
+                   val.document_url,
+                   val.valid,
+                   val.report
             FROM public.document as doc
             LEFT JOIN validation as val ON doc.validation=val.id
             WHERE doc.id = $1;
         `;
-        return module.exports.getFirstRow(sql, [id]);
-    },
+    return getFirstRow(sql, [id]);
+};
 
-    getPublishersWithDocuments: async () => {
-        const sql = `
+const getReportWithoutErrorsForId = async (id) => {
+    const sql = `
+            SELECT val.document_hash as registry_hash,
+                   val.document_id as registry_id, 
+                   doc.name as registry_name,
+                   val.document_url,
+                   val.valid,
+                   ${valReportSummaryOnly}
+            FROM public.document as doc
+            LEFT JOIN validation as val ON doc.validation=val.id
+            WHERE doc.id = $1;
+        `;
+    return getFirstRow(sql, [id]);
+};
+
+const getReportForName = async (name) => {
+    const sql = `
+            SELECT val.document_hash as registry_hash, 
+                   val.document_id as registry_id,
+                   doc.name as registry_name,
+                   val.document_url,
+                   val.valid,
+                   val.report
+            FROM public.document as doc
+            LEFT JOIN validation as val ON doc.validation=val.id
+            WHERE doc.name = $1;
+        `;
+    return getFirstRow(sql, [name]);
+};
+
+const getReportWithoutErrorsForName = async (name) => {
+    const sql = `
+            SELECT val.document_hash as registry_hash, 
+                   val.document_id as registry_id,
+                   doc.name as registry_name,
+                   val.document_url,
+                   val.valid,
+                   ${valReportSummaryOnly}
+            FROM public.document as doc
+            LEFT JOIN validation as val ON doc.validation=val.id
+            WHERE doc.name = $1;
+        `;
+    return getFirstRow(sql, [name]);
+};
+
+const getPublishersWithDocuments = async () => {
+    const sql = `
             SELECT org_id, name, title, state, country_code, package_count, iati_id, black_flag, black_flag_notified
             FROM publisher
             WHERE package_count > 0
         `;
-        return module.exports.query(sql);
-    },
+    return query(sql);
+};
 
-    getPublishersWithBlackFlag: async () => {
-        const sql = `
+const getPublishersWithBlackFlag = async () => {
+    const sql = `
             SELECT org_id, name, description, title, state, image_url, country_code, package_count, iati_id, black_flag, black_flag_notified
             FROM publisher
             WHERE black_flag is not Null
         `;
-        return module.exports.query(sql);
-    },
+    return query(sql);
+};
 
-    getDocumentsForPublisher: async (id) => {
-        const sql = `
+const getDocumentsForPublisher = async (id) => {
+    const sql = `
         SELECT 
             doc.id, 
             doc.hash, 
             doc.url, 
+            doc.name, 
             doc.first_seen, 
             doc.downloaded,
             doc.download_error,
+            doc.file_schema_valid,
             doc.validation,
             doc.regenerate_validation_report,
             doc.publisher,
             doc.modified,
             doc.solrize_end,
-            doc.alv_start,
-            doc.alv_end,
-            doc.alv_error,
+            doc.clean_start,
+            doc.clean_end,
+            doc.clean_error,
             val.created as validation_created, 
             val.valid,
             (SELECT case when val.report is null then null else jsonb_build_object(
@@ -167,11 +215,11 @@ module.exports = {
         WHERE doc.publisher = $1
         ORDER BY url ASC
         `;
-        return module.exports.query(sql, [id]);
-    },
+    return query(sql, [id]);
+};
 
-    getSinglePublisherById: async (id) => {
-        const sql = `
+const getSinglePublisherById = async (id) => {
+    const sql = `
         SELECT
             org_id,
             name,
@@ -185,11 +233,11 @@ module.exports = {
         FROM publisher
         WHERE org_id = $1
         `;
-        return module.exports.query(sql, [id]);
-    },
+    return query(sql, [id]);
+};
 
-    getSinglePublisherByName: async (name) => {
-        const sql = `
+const getSinglePublisherByName = async (name) => {
+    const sql = `
         SELECT
             org_id,
             name,
@@ -203,26 +251,28 @@ module.exports = {
         FROM publisher
         WHERE name = $1
         `;
-        return module.exports.query(sql, [name]);
-    },
+    return query(sql, [name]);
+};
 
-    getSingleDocument: async (id) => {
-        const sql = `
+const getSingleDocumentForId = async (id) => {
+    const sql = `
         SELECT
             doc.id,
             doc.hash,
             doc.url,
+            doc.name,
             doc.first_seen,
             doc.downloaded,
             doc.download_error,
+            doc.file_schema_valid,
             doc.validation,
             doc.regenerate_validation_report,
             doc.publisher,
             doc.modified,
             doc.solrize_end,
-            doc.alv_start,
-            doc.alv_end,
-            doc.alv_error,
+            doc.clean_start,
+            doc.clean_end,
+            doc.clean_error,
             val.created as validation_created,
             val.valid,
             val.report -> 'summary' AS summary
@@ -230,11 +280,39 @@ module.exports = {
         LEFT JOIN validation AS val ON doc.validation = val.id
         WHERE doc.id = $1
         `;
-        return module.exports.query(sql, [id]);
-    },
+    return query(sql, [id]);
+};
+const getSingleDocumentForName = async (name) => {
+    const sql = `
+        SELECT
+            doc.id,
+            doc.hash,
+            doc.url,
+            doc.name,
+            doc.first_seen,
+            doc.downloaded,
+            doc.download_error,
+            doc.file_schema_valid,
+            doc.validation,
+            doc.regenerate_validation_report,
+            doc.publisher,
+            doc.modified,
+            doc.solrize_end,
+            doc.clean_start,
+            doc.clean_end,
+            doc.clean_error,
+            val.created as validation_created,
+            val.valid,
+            val.report -> 'summary' AS summary
+        FROM document as doc
+        LEFT JOIN validation AS val ON doc.validation = val.id
+        WHERE doc.name = $1
+        `;
+    return query(sql, [name]);
+};
 
-    getAdhocValidationSession: async (sessionId) => {
-        const sql = `
+const getAdhocValidationSession = async (sessionId) => {
+    const sql = `
         SELECT 
             guid,
             filename, 
@@ -254,150 +332,72 @@ module.exports = {
         ORDER BY created desc
         `;
 
-        const result = await module.exports.query(sql, [sessionId]);
+    const result = await query(sql, [sessionId]);
 
-        return result;
-    },
+    return result;
+};
 
-    insertAdhocValidation: async (sessionId, filename, guid) => {
-        const sql = `
+const insertAdhocValidation = async (sessionId, filename, guid) => {
+    const sql = `
         INSERT INTO adhoc_validation (session_id, filename, guid) VALUES ($1, $2, $3)
         `;
 
-        const result = await module.exports.query(sql, [sessionId, filename, guid]);
+    const result = await query(sql, [sessionId, filename, guid]);
 
-        return result;
-    },
+    return result;
+};
 
-    updateAdhocValidation: async (guid, sessionId, valid, report, created, errorStatus) => {
-        const sql = `
+const updateAdhocValidation = async (guid, sessionId, valid, report, created, errorStatus) => {
+    const sql = `
        UPDATE adhoc_validation 
        SET valid=$1, report=$2, created=$3, validated=$4, validation_api_error=$5
        WHERE guid=$6 and session_id=$7
         `;
 
-        const now = new Date();
+    const now = new Date();
 
-        const result = await module.exports.query(sql, [
-            valid,
-            JSON.stringify(report),
-            created,
-            now.toISOString(),
-            errorStatus,
-            guid,
-            sessionId,
-        ]);
+    const result = await query(sql, [
+        valid,
+        JSON.stringify(report),
+        created,
+        now.toISOString(),
+        errorStatus,
+        guid,
+        sessionId,
+    ]);
 
-        return result;
-    },
+    return result;
+};
 
-    updateRegenerateValidationForIds: async (ids) => {
-        // Non ALV Docs
-        const sql1 = `
-        UPDATE document
-        SET 
-            regenerate_validation_report = 't'
-        WHERE
-            id = ANY($1)
-            AND alv_end is null
-            AND alv_error is null;
-        `;
-
-        // ALV Docs
-        const sql2 = `
-        UPDATE document
-        SET 
-            regenerate_validation_report = 't',
-            solrize_reindex = 't',
-            lakify_start = null,
-            lakify_end = null,
-            lakify_error = null,
-            flatten_end = null,
-            flatten_start = null,
-            flattened_activities = null,
-            flatten_api_error = null,
-            alv_revalidate = 't',
-            downloaded = null,
-            download_error = null
-        WHERE
-            id = ANY($1)
-            AND alv_end is not null;
-        `;
-
-        // ALV Docs (errored so not in DS)
-        const sql3 = `
-        UPDATE document
-        SET 
-            regenerate_validation_report = 't',
-            alv_revalidate = 't',
-            downloaded = null,
-            download_error = null
-        WHERE
-            id = ANY($1)
-            AND alv_end is null 
-            AND alv_error is not null;
-        `;
-
-        await module.exports.query(sql1, [ids]);
-        await module.exports.query(sql2, [ids]);
-        await module.exports.query(sql3, [ids]);
-    },
-
-    updateRegenerateValidationForAll: async () => {
-        // Non ALV Docs
-        const sql1 = `
+const updateRegenerateValidationForIds = async (ids) => {
+    const sql = `
         UPDATE document
         SET 
             regenerate_validation_report = 't'
         WHERE
             validation is not Null
-            AND alv_end is null
-            AND alv_error is null;
+            AND id = ANY($1)
         `;
 
-        // ALV Docs
-        const sql2 = `
+    await query(sql, [ids]);
+};
+
+const updateRegenerateValidationForAll = async () => {
+    const sql = `
         UPDATE document
         SET 
-            regenerate_validation_report = 't',
-            solrize_reindex = 't',
-            lakify_start = null,
-            lakify_end = null,
-            lakify_error = null,
-            flatten_end = null,
-            flatten_start = null,
-            flattened_activities = null,
-            flatten_api_error = null,
-            alv_revalidate = 't',
-            downloaded = null,
-            download_error = null
-        WHERE
-            validation is not Null
-            AND alv_end is not null;
+            regenerate_validation_report = 't'
+        WHERE validation is not Null
         `;
 
-        // ALV Docs (errored so not in DS)
-        const sql3 = `
-        UPDATE document
-        SET 
-            regenerate_validation_report = 't',
-            alv_revalidate = 't',
-            downloaded = null,
-            download_error = null
-        WHERE
-            validation is not Null
-            AND alv_end is null 
-            AND alv_error is not null;
-        `;
+    const result = await query(sql);
 
-        await module.exports.query(sql1);
-        await module.exports.query(sql2);
-        await module.exports.query(sql3);
-    },
+    return result;
+};
 
-    getSummaryPrecalcStats: async (date, publisher) => {
-        if (publisher) {
-            const sql = `
+const getSummaryPrecalcStats = async (date, publisher) => {
+    if (publisher) {
+        const sql = `
             SELECT
                 T1.publisher_name,
                 SUM( (T1.report -> 'summary' ->> 'critical') :: INTEGER) as critical,
@@ -415,10 +415,10 @@ module.exports = {
             GROUP BY T1.publisher_name;
             `;
 
-            const result = await module.exports.query(sql, [date, publisher]);
-            return result;
-        }
-        const sql = `
+        const result = await query(sql, [date, publisher]);
+        return result;
+    }
+    const sql = `
             SELECT
                 T1.publisher_name,
                 SUM( (T1.report -> 'summary' ->> 'critical') :: INTEGER) as critical,
@@ -436,13 +436,13 @@ module.exports = {
             GROUP BY T1.publisher_name;
             `;
 
-        const result = await module.exports.query(sql, [date]);
-        return result;
-    },
+    const result = await query(sql, [date]);
+    return result;
+};
 
-    getSummaryAggregateStats: async (date, publisher) => {
-        if (publisher) {
-            const sql = `
+const getSummaryAggregateStats = async (date, publisher) => {
+    if (publisher) {
+        const sql = `
                 SELECT
                     T1.publisher_name,
                     arr3.item_object -> 'severity' AS severity,
@@ -462,10 +462,10 @@ module.exports = {
                 ) GROUP BY T1.publisher_name, severity;
             `;
 
-            const result = await module.exports.query(sql, [date, publisher]);
-            return result;
-        }
-        const sql = `
+        const result = await query(sql, [date, publisher]);
+        return result;
+    }
+    const sql = `
             SELECT
                 T1.publisher_name,
                 arr3.item_object -> 'severity' AS severity,
@@ -485,12 +485,12 @@ module.exports = {
             ) GROUP BY T1.publisher_name, severity;
         `;
 
-        const result = await module.exports.query(sql, [date]);
-        return result;
-    },
+    const result = await query(sql, [date]);
+    return result;
+};
 
-    getMessageDateStats: async (date) => {
-        const sql = `
+const getMessageDateStats = async (date) => {
+    const sql = `
             SELECT
                 T1.publisher_name,
                 arr3.item_object -> 'id' AS error_id,
@@ -513,12 +513,12 @@ module.exports = {
             ) GROUP BY T1.publisher_name, error_id, message, severity, category;
         `;
 
-        const result = await module.exports.query(sql, [date]);
-        return result;
-    },
+    const result = await query(sql, [date]);
+    return result;
+};
 
-    getMessagePublisherStats: async (date, publisher) => {
-        const sql = `
+const getMessagePublisherStats = async (date, publisher) => {
+    const sql = `
             SELECT
                 T1.publisher_name,
                 arr3.item_object -> 'id' AS error_id,
@@ -541,7 +541,36 @@ module.exports = {
             ) GROUP BY T1.publisher_name, error_id, message, severity, category;
         `;
 
-        const result = await module.exports.query(sql, [date, publisher]);
-        return result;
-    },
+    const result = await query(sql, [date, publisher]);
+    return result;
+};
+
+export {
+    query,
+    getFirstRow,
+    getReportForUrl,
+    getReportWithoutErrorsForUrl,
+    getReportForTestfile,
+    getReportForHash,
+    getReportWithoutErrorsForHash,
+    getReportForId,
+    getReportWithoutErrorsForId,
+    getReportForName,
+    getReportWithoutErrorsForName,
+    getPublishersWithDocuments,
+    getPublishersWithBlackFlag,
+    getDocumentsForPublisher,
+    getSinglePublisherById,
+    getSinglePublisherByName,
+    getSingleDocumentForId,
+    getSingleDocumentForName,
+    getAdhocValidationSession,
+    insertAdhocValidation,
+    updateAdhocValidation,
+    updateRegenerateValidationForIds,
+    updateRegenerateValidationForAll,
+    getSummaryPrecalcStats,
+    getSummaryAggregateStats,
+    getMessageDateStats,
+    getMessagePublisherStats,
 };
