@@ -1,6 +1,6 @@
-const db = require('../database/db');
+import { getMessagePublisherStats } from '../database/db.js';
 
-module.exports = async (context, req) => {
+export default async function pubGetStatsPublisher(context, req) {
     const date = req.query.date ? req.query.date : '9999-01-01';
     const { publisher } = req.params;
     const { format } = req.query;
@@ -23,7 +23,7 @@ module.exports = async (context, req) => {
     try {
         let result = null;
 
-        result = await db.getMessagePublisherStats(date, publisher);
+        result = await getMessagePublisherStats(date, publisher);
 
         if (format === 'csv') {
             const csvString = [
@@ -74,4 +74,4 @@ module.exports = async (context, req) => {
             body: JSON.stringify(e),
         };
     }
-};
+}
